DERBY-6000: Fix misc issues raised by JDBC 4.2 buddy testing.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1450422 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/functionTests/tests/jdbc4/UnsupportedVetter.java
@@ -814,7 +814,9 @@ public class UnsupportedVetter	extends BaseJDBCTestCase
             return new Short((short) 0);
         }
         if (type == Integer.TYPE) {
-            return new Integer(0);
+            // use 1 instead of 0. 0 = java.sql.Types.NULL and this causes
+            // various type-aware methods to raise SQLFeatureNotSupportedException
+            return new Integer(1);
         }
         if (type == Long.TYPE) {
             return new Long(0L);
